Implement approval and rejection by a superuser of a Activation
