+++ main/test/activation_fixtures.json
