+++ main/views.py
@@ -1,1 +1,5 @@
-#
+from django.views.generic.base import TemplateView
+
+class IndexView(TemplateView):
+
+    template_name = "index.html"
